rename "implicit conversion test" to "type declaration bounds test"
