--- framework/src/test/java/org/checkerframework/framework/testchecker/implicitconversion/ImplicitConversionTestChecker.java
@@ -1,5 +1,0 @@
-package org.checkerframework.framework.testchecker.implicitconversion;
-
-import org.checkerframework.common.basetype.BaseTypeChecker;
-
-public class ImplicitConversionTestChecker extends BaseTypeChecker {}
