+++ framework/src/test/java/org/checkerframework/framework/testchecker/typedeclbounds/TypeDeclBoundsAnnotatedTypeFactory.java
@@ -1,18 +1,18 @@
-package org.checkerframework.framework.testchecker.implicitconversion;
+package org.checkerframework.framework.testchecker.typedeclbounds;
 
 import org.checkerframework.common.basetype.BaseAnnotatedTypeFactory;
 import org.checkerframework.common.basetype.BaseTypeChecker;
-import org.checkerframework.framework.testchecker.implicitconversion.quals.Bottom;
-import org.checkerframework.framework.testchecker.implicitconversion.quals.Top;
+import org.checkerframework.framework.testchecker.typedeclbounds.quals.Bottom;
+import org.checkerframework.framework.testchecker.typedeclbounds.quals.Top;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 
-public class ImplicitConversionTestAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
+public class TypeDeclBoundsAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
 
-    public ImplicitConversionTestAnnotatedTypeFactory(BaseTypeChecker checker) {
+    public TypeDeclBoundsAnnotatedTypeFactory(BaseTypeChecker checker) {
         super(checker);
         postInit();
     }
