+++ framework/src/test/java/org/checkerframework/framework/testchecker/typedeclbounds/TypeDeclBoundsChecker.java
@@ -1,0 +1,5 @@
+package org.checkerframework.framework.testchecker.typedeclbounds;
+
+import org.checkerframework.common.basetype.BaseTypeChecker;
+
+public class TypeDeclBoundsChecker extends BaseTypeChecker {}
